+++ infrastructure/lib/constants.ts
@@ -11,7 +11,6 @@ export const CONFIG = {
     },
     email: {
       sender: "no-reply@dev.bjornmelin.io",
-      recipient: "dev-contact@bjornmelin.io",
       allowedOrigins: [
         "https://dev.bjornmelin.io",
         "https://www.dev.bjornmelin.io",
@@ -31,7 +30,6 @@ export const CONFIG = {
     },
     email: {
       sender: "no-reply@bjornmelin.io",
-      recipient: "bjornmelin16@gmail.com",
       allowedOrigins: [
         "https://bjornmelin.io",
         "https://www.bjornmelin.io",